+++ include/unity/scopes/internal/LocationImpl.h
@@ -22,6 +22,8 @@
 #include <unity/scopes/Location.h>
 
 #include <unordered_set>
+
+#include <unity/scopes/internal/max_align_clang_bug.h>  // TODO: remove this once clang 3.5.2 is released
 #include <boost/optional.hpp>
 
 namespace unity