+++ include/unity/scopes/internal/ResultReplyObject.h
@@ -37,10 +37,8 @@ class RuntimeImpl;
 class ResultReplyObject : public ReplyObject
 {
 public:
-    ResultReplyObject(SearchListenerBase::SPtr const& receiver,
-                      RuntimeImpl const* runtime,
-                      std::string const& scope_id,
-                      int cardinality);
+    ResultReplyObject(SearchListenerBase::SPtr const& receiver, RuntimeImpl const* runtime,
+                      std::string const& scope_id, int cardinality, bool dont_reap = false);
     virtual ~ResultReplyObject();
 
     virtual bool process_data(VariantMap const& data) override;