+++ src/scopes/internal/RegistryObject.cpp
@@ -24,6 +24,7 @@
 #include <unity/UnityExceptions.h>
 #include <unity/util/ResourcePtr.h>
 
+#include <unity/scopes/internal/max_align_clang_bug.h>  // TODO: remove this once clang 3.5.2 is released
 #include <boost/algorithm/string.hpp>
 #include <boost/filesystem.hpp>
 #include <core/posix/child_process.h>
@@ -260,7 +261,16 @@ bool RegistryObject::remove_local_scope(std::string const& scope_id)
                                               "with empty id");
     }
 
-    lock_guard<decltype(mutex_)> lock(mutex_);
+    unique_lock<decltype(mutex_)> lock(mutex_);
+
+    auto proc_it = scope_processes_.find(scope_id);
+    if (proc_it != scope_processes_.end())
+    {
+        // Kill process after unlocking, so we can handle on_process_death
+        lock.unlock();
+        proc_it->second.kill();
+        lock.lock();
+    }
 
     scope_processes_.erase(scope_id);
 
@@ -330,12 +340,12 @@ void RegistryObject::on_state_received(std::string const& scope_id, StateReceive
 RegistryObject::ScopeProcess::ScopeProcess(ScopeExecData exec_data, MWPublisher::SPtr publisher)
     : exec_data_(exec_data)
     , reg_publisher_(publisher)
+    , manually_started_(false)
 {
 }
 
 RegistryObject::ScopeProcess::ScopeProcess(ScopeProcess const& other)
-    : exec_data_(other.exec_data_)
-    , reg_publisher_(other.reg_publisher_)
+    : ScopeProcess(other.exec_data_, other.reg_publisher_)
 {
 }
 
@@ -443,8 +453,8 @@ void RegistryObject::ScopeProcess::exec(
         env["LD_LIBRARY_PATH"] = scope_ld_lib_path;  // Overwrite any LD_LIBRARY_PATH entry that may already be there.
 
         process_ = executor->exec(program, argv, env,
-                                     core::posix::StandardStream::stdin | core::posix::StandardStream::stdout,
-                                     exec_data_.confinement_profile);
+                                  core::posix::StandardStream::stdin | core::posix::StandardStream::stdout,
+                                  exec_data_.confinement_profile);
         if (process_.pid() <= 0)
         {
             clear_handle_unlocked();
@@ -525,8 +535,7 @@ void RegistryObject::ScopeProcess::update_state_unlocked(ProcessState new_state)
             cout << "RegistryObject::ScopeProcess: Process for scope: \"" << exec_data_.scope_id
                  << "\" started manually" << endl;
 
-            // Don't update state_, treat this scope as not running if a locate() is requested
-            return;
+            manually_started_ = true;
         }
     }
     else if (new_state == Stopped)
@@ -543,7 +552,7 @@ void RegistryObject::ScopeProcess::update_state_unlocked(ProcessState new_state)
                  << "\" closed unexpectedly. Either the process crashed or was killed forcefully." << endl;
         }
     }
-    else if (new_state == Stopping && state_ != Running)
+    else if (new_state == Stopping && manually_started_)
     {
         if (reg_publisher_)
         {
@@ -554,8 +563,8 @@ void RegistryObject::ScopeProcess::update_state_unlocked(ProcessState new_state)
         cout << "RegistryObject::ScopeProcess: Manually started process for scope: \""
              << exec_data_.scope_id << "\" exited" << endl;
 
-        // Don't update state_, treat this scope as not running if a locate() is requested
-        return;
+        new_state = Stopped;
+        manually_started_ = false;
     }
     state_ = new_state;
     state_change_cond_.notify_all();